agents/adk_agent/pyproject.toml
[project]
name = "adk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "google-adk>=1.5.0",
    "google-genai>=1.23.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]
